pyproject.toml
[project]
name = "eazyliving"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.14"
 
dependencies = [
    "bcrypt>=5.0.0",
    "sqladmin>=0.18.0",
    "fastapi[standard]",
    "httpx>=0.28.1",
    "itsdangerous>=2.0.0",
    "passlib>=1.7.4",
    "psycopg2>=2.9.11",
    "pytest>=9.0.3",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.1.0",
    "python-jose[cryptography]>=3.5.0",
    "uvicorn",
]

entrypoint = "gateway.main:app"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
asyncio_default_test_loop_scope = "session"
